requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.4
paho-mqtt==2.1.0
python-dotenv==1.0.1
Jinja2==3.1.5
pytest==8.3.4
httpx==0.28.1
